requirements.txt
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
bleach==6.2.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.43.0
colorama==0.4.6
cookiecutter==2.6.0
diff-match-patch==20241021
dj-database-url==2.3.0
dj3-cloudinary-storage==0.0.6
Django==5.1.7
django-allauth==65.5.0
django-baton==4.2.2
django-bootstrap5==25.1
django-browser-reload==1.18.0
django-import-export==4.3.7
django-jinja==2.11.0
django-unfold==0.52.0
gunicorn==23.0.0
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
pip-review==1.3.0
psycopg2==2.9.10
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
text-unidecode==1.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
webencodings==0.5.1
whitenoise==6.9.0
